feat: Add a nix flake providing a devShell and essential-server package (#48)
* feat: Add a nix flake providing a devShell and essential-server package

This enables using the `nix develop` command to enter a shell with all
the tools and system deps required to work on `essential-server`. Was
going to whip up a shell for myself, but thought I might as well do a
proper one for the repo!

This flake may come in handy as a foundation when it comes time for
working on deployment too.

Currently I've updated CI just to check the validity of the flake and
the nix code formatting, but I haven't plugged it into our actual CI
to avoid adding any extra friction. That said, we may want to consider
integrating it into the CI more thoroughly in the future as it might
help to catch cases where we inadvertently introduce requirements on new
system deps, and will likely make deploy-related jobs easier.

As a sidenote, I'm a little surprised CI works currently without
declaring the openssl and pkg-config system deps, but I guess github
must ship those by default in their containers as they're such common
dependencies or something.

* fix: Provide SSH access to Nix jobs

* fix: Temporarily disable Nix until private SSH access is solved

See issue #48.

---------

Co-authored-by: mitchmindtree <mitchell.nordine@essentialcontributions.com>