feat: Add state read VM (#61)
* feat: Add state read VM (WIP)

* Complete remaining state-read VM operation implementations

* Rm unnecessary Continue type alias

* Add Future with gas-bounded yielding WIP

* ExecFuture takes Vm by &mut, add BytecodeMapped, begin testing

Also introduces the `OpGasCost` and `OpAccess` traits for greater
execution flexibility. `OpAccess` removes the need for differing
`Future` implementations for the &[Op] and BytecodeMapped operation
execution.

* Improve From Infallible error implementation

* Add a lazy bytecode execution method to the Vm

* Upstream parse_op as Op::from_bytes

* Upstream bool_from_word function

* Add Vm::into_state_slots. Improve crate root and method docs.

* Don't expose step_op fns

Used incorrectly, these can leave the VM in an inconsistent state.
They are implementation details in the execution implementation. We can
consider providing `step_op` methods on `Vm` that reliably leave the
`Vm` in a consistent state.

* Rm unnecessary comment from Memory::alloc

* Rm Op::Stack error variant in favour of Op::Constraint

* Deny missing_docs, unsafe_code

* Add control flow operation tests

* Add first basic state read test

* Add tests for all memory ops

* Add memory error case tests

* Add a test for invalid jump condition

* Add state read op tests, consolidate State impl

Also adds a test that ensures each of the execution methods behave the
same.

* Add detailed pre-post state and constraint check test. Run clippy.

* Don't expose some Op-related implementation details. Improve docs.

* Remove old debugging printlns

* Fix leftover typo in the docs

* Rm unnecessary boxing of ExecFuture. Requires OpAccess is Unpin.

* Move state read VM tests into dedicated tests dir

* Switch back to essential-base main branch

* Add an associated Future to the StateRead trait

This allows for removing the `PendingOp` future allocation within the
`ExecFuture` `poll` implementation. This in turn allows `StateRead`
implementers to decide on whether the cost of `Box`ing the state read
future is worth it.

The `test/util.rs` `State` type now demonstrates how the new associated
`Future` type allows implementers to avoid allocations entirely with its
simple in-memory implementation using `std::future::Ready`.

* Fix incomplete doc on StateReadFuture

* Improve BytecodeMappedLazy op_access implementation

* Use `size_hint` in `FromIterator` for BytecodeMapped. Move tests.

Also adds a conversion from a `Vec<u8>` that maps an existing, owned Vec
of bytecoode.

* Require that programs always end with `Halt`. Add PcOutOfRange tests.

* Return an error on program counter overflow

* Improve `yield_per_op` behaviour to actually test the yield count