refactor: Map state read bytecode ahead of execution
This avoids building the operation mapping twice in favour of mapping
ahead of execution and sharing the mapping between both pre and post
state slot reads.

Also updates the `essential-base` crates to make use of the new support
for mapping over an existing bytecode slice.

Note that this does slightly change up the performance trade-offs, e.g.
in the case that a state read contains a lot of bytecode but execution
would fail immediately, this new approach may be more costly.

However, the same could be said for the inverse, e.g. if the state
read does a lot of heavy compute but the end of the bytecode contains
an invalid operation, this would avoid doing any unnecessary execution
in the first place. I think this change is worth it, as mapping the
bytecode is predictably much faster than execution.